tests with lists instead of file
