+++ opendse-io/src/integration-test/java/net/sf/opendse/io/ReaderTGFFIntegrationTest.java
@@ -9,18 +9,18 @@ import net.sf.opendse.model.Dependency;
 import net.sf.opendse.model.Mappings;
 import net.sf.opendse.model.Resource;
 import net.sf.opendse.model.ResourceTypes;
+import net.sf.opendse.model.SpecificationTypeBased;
 import net.sf.opendse.model.Task;
-import net.sf.opendse.model.TypeBasedSpecification;
 
-public class TGFFReaderIntegrationTest {
+public class ReaderTGFFIntegrationTest {
 
-	private final String testFile = "specs/opendse_example.tgff";
+	private final String testFile = "specs/tgff/opendse_example.tgff";
 
 	@Test
 	public void applicationTest() {
 
-		TGFFReader reader = new TGFFReader();
-		TypeBasedSpecification spec = reader.read(testFile);
+		ReaderTGFF reader = new ReaderTGFF();
+		SpecificationTypeBased spec = reader.read(testFile);
 		Application<Task, Dependency> application = spec.getApplication();
 
 		Assert.assertNotNull(application);
@@ -45,32 +45,32 @@ public class TGFFReaderIntegrationTest {
 	@Test
 	public void compTaskAttributeTest() {
 
-		TGFFReader reader = new TGFFReader();
-		TypeBasedSpecification spec = reader.read(testFile);
+		ReaderTGFF reader = new ReaderTGFF();
+		SpecificationTypeBased spec = reader.read(testFile);
 		Task task = (spec.getApplication()).getVertex("t0_3_0");
 
-		Assert.assertEquals(300.0, task.getAttribute(TGFFReader.PERIOD));
-		Assert.assertEquals(200.0, task.getAttribute(TGFFReader.HARD_DEADLINE));
-		Assert.assertEquals("2", task.getAttribute(TGFFReader.TGFF_TYPE));
+		Assert.assertEquals(300.0, task.getAttribute(ReaderTGFF.PERIOD));
+		Assert.assertEquals(200.0, task.getAttribute(ReaderTGFF.SOFT_DEADLINE));
+		Assert.assertEquals("2", task.getAttribute(ReaderTGFF.TGFF_TYPE));
 	}
 
 	@Test
 	public void commTaskAttributeTest() {
 
-		TGFFReader reader = new TGFFReader();
-		TypeBasedSpecification spec = reader.read(testFile);
+		ReaderTGFF reader = new ReaderTGFF();
+		SpecificationTypeBased spec = reader.read(testFile);
 		Communication communication = (Communication) (spec.getApplication()).getVertex("a0_0");
 
-		Assert.assertEquals(300.0, communication.getAttribute(TGFFReader.PERIOD));
-		Assert.assertEquals("0", communication.getAttribute(TGFFReader.TGFF_TYPE));
-		Assert.assertEquals(5.0, communication.getAttribute(TGFFReader.MSG_SIZE));
+		Assert.assertEquals(300.0, communication.getAttribute(ReaderTGFF.PERIOD));
+		Assert.assertEquals("0", communication.getAttribute(ReaderTGFF.TGFF_TYPE));
+		Assert.assertEquals(5.0, communication.getAttribute(ReaderTGFF.MSG_SIZE));
 	}
 
 	@Test
 	public void resourceTypesTest() {
 
-		TGFFReader reader = new TGFFReader();
-		TypeBasedSpecification spec = reader.read(testFile);
+		ReaderTGFF reader = new ReaderTGFF();
+		SpecificationTypeBased spec = reader.read(testFile);
 		ResourceTypes<?> resourceTypes = spec.getResourceTypes();
 
 		for (int i = 0; i <= 2; i++) {
@@ -83,8 +83,8 @@ public class TGFFReaderIntegrationTest {
 	@Test
 	public void resourceAttributeTest() {
 
-		TGFFReader reader = new TGFFReader();
-		TypeBasedSpecification spec = reader.read(testFile);
+		ReaderTGFF reader = new ReaderTGFF();
+		SpecificationTypeBased spec = reader.read(testFile);
 		Resource resource = (spec.getResourceTypes()).get("r0");
 
 		Assert.assertNotNull(resource);
@@ -96,8 +96,8 @@ public class TGFFReaderIntegrationTest {
 	@Test
 	public void mappingsTest() {
 
-		TGFFReader reader = new TGFFReader();
-		TypeBasedSpecification spec = reader.read(testFile);
+		ReaderTGFF reader = new ReaderTGFF();
+		SpecificationTypeBased spec = reader.read(testFile);
 		Mappings<Task, Resource> mappings = spec.getMappings();
 
 		Task task = (spec.getApplication()).getVertex("t0_0_0");
